+++ tests/utilities.js
@@ -1,0 +1,7 @@
+let wrapper = require('../wrapper.js');
+
+global.ResilientWrapper = wrapper.ResilientWrapper;
+
+global.game = {
+	ready: true
+};